Generate coverage information to guide fuzzer
This generates coverage information from the fuzz generator
which is used to guide the fuzzer. The coverage should be
roughly the same as what would have been generated for the
parent process (same for number of executions per line) had
it been possible to fuzz multithreaded code with tsan enabled
directly.